rag_application_br_music/backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
openai
requests